docs: Resolve {{version}} placeholders in generated docs at build time
- Replace {{version}} placeholders in network-schema.gpkg on RTD builds
- Update logic for replacing {{version}} placeholders in network-schema.json on RTD builds
- Replace {{version}} placeholders on local builds
- Move logic for generating GeoPackage table definitions from manage.py to conf.py (Sphinx build)
