--- schema/geopackage/table_definitions/contracts.csv
@@ -1,10 +1,0 @@
-Name,Type,Constraints,Title,Description
-id,INTEGER,PK,,
-ofds_id,TEXT,,Identifier,A locally-unique identifier for this contract.
-title,TEXT,,Title,A title for this contract.
-description,TEXT,,Description,A description for this contract.
-type,INTEGER,FK,Type,"The type of this contract, from the open [contractType codelist](https://standard.ofds.info/en/{{version}}/reference/codelists#contracttype)."
-value__amount,REAL,,Value Amount,The value of this contract. The amount of this value.
-value__currency,TEXT,,Value Currency,"The value of this contract. The currency of this value, from the closed [currency codelist](https://standard.ofds.info/en/{{version}}/reference/codelists#currency)."
-dateSigned,TEXT,,Date signed,The date this contract was signed.
-network_id,INTEGER,FK,Network ID,
