--- schema/geopackage/table_definitions/contracts_documents.csv
@@ -1,8 +1,0 @@
-Name,Type,Constraints,Title,Description
-id,INTEGER,PK,,
-title,TEXT,,Title,A title for this document.
-description,TEXT,,Description,"A description of this document. Descriptions should not exceed 250 words. If the document is not accessible online, the description should describe how to access a copy of the document."
-url,TEXT,,URL,A web address for accessing this document.
-format,INTEGER,FK,Format,"The format of this document, from the open [mediaType codelist](https://standard.ofds.info/en/{{version}}/reference/codelists#mediatype)."
-network_id,INTEGER,FK,Network ID,
-contract_id,INTEGER,FK,Contract ID,
