--- schema/geopackage/table_definitions/networks.csv
@@ -1,16 +1,0 @@
-Name,Type,Constraints,Title,Description
-id,INTEGER,PK,,
-ofds_id,TEXT,,UUID,"A universally unique identifier for this network, as defined by [RFC 4122](https://datatracker.ietf.org/doc/html/rfc4122). For more information, see the [identifier reference](https://standard.ofds.info/en/{{version}}/reference/identifiers)."
-identifier,TEXT,,Identifier,A local identifier for this network in an internal system.
-name,TEXT,,Network name,A name for this network.
-website,TEXT,,Website,The URL of the website for this network.
-publisher__name,TEXT,,Publisher Name,The publisher of this network. A name for this organisation.
-publisher__identifier__id,TEXT,,Publisher Organisation identifier Identifier,The publisher of this network. An identifier for this organisation. The identifier assigned to the organisation in the register identified in `.scheme`.
-publisher__identifier__scheme,INTEGER,FK,Publisher Organisation identifier Scheme,"The publisher of this network. An identifier for this organisation. The register from which the identifier in `.id` is drawn, from the open [organisationIdentifierScheme codelist](https://standard.ofds.info/en/{{version}}/reference/codelists#organisationidentifierscheme)."
-publisher__identifier__legalName,TEXT,,Publisher Organisation identifier Legal name,The publisher of this network. An identifier for this organisation. The legally registered name of the organisation
-publisher__identifier__uri,TEXT,,Publisher Organisation identifier URI,"The publisher of this network. An identifier for this organisation. A canonical URI for this identifier, such as those provided by Open Corporates. Do not use this field to provide the website of the organisation: instead, use `Organisation.website`."
-publicationDate,TEXT,,Publication date,The date when this network was published.
-collectionDate,TEXT,,Collection date,"The date when the location data was collected. If a dataset was produced by digitising a map, the date that the data for the map was collected."
-accuracy,REAL,,Accuracy,"The horizontal uncertainty, in metres, of the coordinates in this dataset relative to the datum of the coordinate reference system specified in `crs`. Further details about the accuracy of coordinates in this dataset can be provided in `.accuracyDetails`."
-accuracyDetails,TEXT,,Accuracy details,"Further details about the accuracy specified in `accuracy`. For example, the confidence level of the accuracy measurement, the methodology used to calculate the accuracy, or a local classification of accuracy."
-language,INTEGER,FK,Language,"The default language used in text attributes,from the open [language codelist](https://standard.ofds.info/en/{{version}}/reference/codelists#language). A [BCP47 language tag](https://www.w3.org/International/articles/language-tags/) is allowed, if there is a user need for the additional information."
