--- schema/geopackage/table_definitions/networks_fks.csv
@@ -1,3 +1,0 @@
-Column,References
-language,FK(`codelist_open_language`.`id`)
-publisher__identifier__scheme,FK(`codelist_open_organisationIdentifierScheme`.`id`)
