--- schema/geopackage/table_definitions/nodes_fks.csv
@@ -1,6 +1,0 @@
-Column,References
-network_id,FK(`networks`.`id`)
-transmissionMediumOwner,FK(`organisations`.`id`)
-supportingInfrastructure__owner,FK(`organisations`.`id`)
-phase,FK(`phases`.`id`)
-supportingInfrastructure__type,FK(`codelist_open_nodeSupportingInfrastructure`.`id`)
