--- schema/geopackage/table_definitions/organisations.csv
@@ -1,13 +1,0 @@
-Name,Type,Constraints,Title,Description
-id,INTEGER,PK,,
-ofds_id,TEXT,,Local identifier,A locally-unique identifier for this organisation.
-name,TEXT,,Name,A name for this organisation. The name should be unique amongst the organisations belonging to a network.
-identifier__id,TEXT,,Organisation identifier Identifier,An identifier for this organisation. The identifier assigned to the organisation in the register identified in `.scheme`.
-identifier__scheme,INTEGER,FK,Organisation identifier Scheme,"An identifier for this organisation. The register from which the identifier in `.id` is drawn, from the open [organisationIdentifierScheme codelist](https://standard.ofds.info/en/{{version}}/reference/codelists#organisationidentifierscheme)."
-identifier__legalName,TEXT,,Organisation identifier Legal name,An identifier for this organisation. The legally registered name of the organisation
-identifier__uri,TEXT,,Organisation identifier URI,"An identifier for this organisation. A canonical URI for this identifier, such as those provided by Open Corporates. Do not use this field to provide the website of the organisation: instead, use `Organisation.website`."
-country,TEXT,,Country,"The country in which this organisation is legally registered, from the closed [country codelist](https://standard.ofds.info/en/{{version}}/reference/codelists#country)."
-roleDetails,TEXT,,Role details,Further details about this organisation's roles in the network
-website,TEXT,,Website,The URL of the website for this organisation.
-logo,TEXT,,Logo,The URL of the logo for this organisation.
-network_id,INTEGER,FK,Network ID,
