--- schema/geopackage/table_definitions/organisations_fks.csv
@@ -1,3 +1,0 @@
-Column,References
-network_id,FK(`networks`.`id`)
-identifier__scheme,FK(`codelist_open_organisationIdentifierScheme`.`id`)
