--- schema/geopackage/table_definitions/phases_fks.csv
@@ -1,2 +1,0 @@
-Column,References
-network_id,FK(`networks`.`id`)
