--- schema/geopackage/table_definitions/spans.csv
@@ -1,28 +1,0 @@
-Name,Type,Constraints,Title,Description
-id,INTEGER,PK,,
-geom,BLOB,,,
-ofds_id,TEXT,,Identifier,A locally-unique identifier for this span.
-name,TEXT,,Name,A name for this span. The name should be unique amongst the spans belonging to a network.
-phase,INTEGER,FK,Phase,The phase to which this span belongs.
-status,TEXT,,Status,"The status of the network infrastructure for this span, from the closed [spanStatus codelist](https://standard.ofds.info/en/{{version}}/reference/codelists#spanstatus)."
-readyForServiceDate,TEXT,,Ready for service date,"The date this span was ready to carry traffic. For spans with multiple network providers, the earliest of the dates at which each provider's network was ready to carry traffic on this span."
-start,INTEGER,FK,Start,"The identifier of one of two nodes that this span connects. If the span is directed, it represents the starting point of the span."
-end,INTEGER,FK,End,"The identifier of one of two nodes that this span connects. If the span is directed, it represents the end point of the span."
-directed,TEXT,,Directed,"Whether this span is directed. If the span is directed, the span begins at the node referenced in `.start` and ends at the node referenced in `.end`. Otherwise, the span connects the nodes referenced in `.start` and `.end` with no direction implied."
-transmissionMediumOwner,INTEGER,FK,Owner,"The organisation that owns the transmission media for this span, i.e. the fibre cables."
-supplier,INTEGER,FK,Supplier,The organisation responsible for installing the cable for this span.
-supportingInfrastructure__type,INTEGER,FK,Supporting infrastructure Type,"The infrastructure that supports this span. The type of supporting infrastructure, from the open [spanSupportingInfrastructure codelist](https://standard.ofds.info/en/{{version}}/reference/codelists#spansupportinginfrastructure)."
-supportingInfrastructure__description,TEXT,,Supporting infrastructure Description,The infrastructure that supports this span. A description of the supporting infrastructure for this span.
-supportingInfrastructure__owner,INTEGER,FK,Supporting infrastructure Owner,The infrastructure that supports this span. The organisation that owns the supporting infrastructure.
-supportingInfrastructure__spareCapacity,TEXT,,Supporting infrastructure Spare capacity,The infrastructure that supports this span. Whether the supporting infrastructure has spare capacity for the installation of additional fibre cables.
-codeployment,INTEGER,FK,Codeployment,The type of utility or transport infrastrucutre with which this span is codeployed.
-cableType,INTEGER,FK,Cable type,The type of cable used in this span.
-darkFibre,TEXT,,Dark fibre availability,Whether access to dark fibre is available on this span.
-fibreType,TEXT,,Fibre type,"The type of fibre used in this span, from the closed [fibreType codelist](https://standard.ofds.info/en/{{version}}/reference/codelists#fibretype). Further details of the span's fibre type can be provided in `.fibreTypeDetails`."
-fibreTypeDetails__fibreSubtype,TEXT,,Fibre type details Subtype,"Further details about this span's fibre type. The sub-category of this span's fibre type. For example, G.652.B."
-fibreTypeDetails__description,TEXT,,Fibre type details Description,Further details about this span's fibre type. A description of this span's fibre type.
-fibreCount,INTEGER,,Fibre count,The number of individual optical fibres in this span.
-fibreLength,REAL,,Fibre length,"The physical length in kilometres, of fibre optic cable used in this span."
-capacity,REAL,,Equipped capacity,"The transmission rate, or throughput, of this span, expressed in Gbit/sec (Gbps). The equipped capacity is the total capacity of the circuits (e.g. E1, DS3, STM-1 etc.) which have been activated in the network transmission equipment of the span. For spans with multiple network providers, the sum of the capacity of each provider's network on this span. Further details of this span's capacity can be provided in `.capacityDetails`. For more information, see [equipped capacity](https://standard.ofds.info/en/{{version}}/reference/data_model/#equipped-capacity)."
-capacityDetails__description,TEXT,,Capacity details Description,Further details about this span's capacity. A description of this span's capacity.
-network_id,INTEGER,FK,Network ID,
