--- schema/geopackage/table_definitions/spans_fks.csv
@@ -1,11 +1,0 @@
-Column,References
-network_id,FK(`networks`.`id`)
-supportingInfrastructure__owner,FK(`organisations`.`id`)
-supplier,FK(`organisations`.`id`)
-transmissionMediumOwner,FK(`organisations`.`id`)
-end,FK(`nodes`.`id`)
-start,FK(`nodes`.`id`)
-phase,FK(`phases`.`id`)
-cableType,FK(`codelist_open_cableType`.`id`)
-codeployment,FK(`codelist_open_codeployment`.`id`)
-supportingInfrastructure__type,FK(`codelist_open_spanSupportingInfrastructure`.`id`)
