--- schema/geopackage/table_definitions/wayleaves.csv
@@ -1,11 +1,0 @@
-Name,Type,Constraints,Title,Description
-id,INTEGER,PK,,
-ofds_id,TEXT,,Identifier,A locally-unique identifier for this wayleave.
-grantor,INTEGER,FK,Grantor,"The entity that grants a network operator the legal right to install, maintain, or operate telecommunications infrastructure on, over, or under property or assets."
-yearSigned,INTEGER,,Year signed,The year in which this wayleave was signed.
-term__indefinite,TEXT,,Term Indefinite,"Information about the duration of the wayleave. Whether the term of wayleave is indefinite, i.e. without an end date."
-term__years,INTEGER,,Term Years,"Information about the duration of the wayleave. If the wayleave's term is definite, its duration in years."
-cost__recurring,TEXT,,Cost Recurring,Information about the cost of the wayleave. Whether this is a recurring cost.
-cost__perMetre__amount,REAL,,Cost Cost per metre Amount,"Information about the cost of the wayleave. The average cost per metre for this wayleave. Wayleave costs can include a mix of per length and per item fees, e.g. a fee per metre of cable and a fee per pole. The average cost per metre should be calculated by dividing the total cost of all fees by the total length in metres of cable covered by the wayleave. For recurring fees, the average cost per metre should be annualised. The amount of this value."
-cost__perMetre__currency,TEXT,,Cost Cost per metre Currency,"Information about the cost of the wayleave. The average cost per metre for this wayleave. Wayleave costs can include a mix of per length and per item fees, e.g. a fee per metre of cable and a fee per pole. The average cost per metre should be calculated by dividing the total cost of all fees by the total length in metres of cable covered by the wayleave. For recurring fees, the average cost per metre should be annualised. The currency of this value, from the closed [currency codelist](https://standard.ofds.info/en/{{version}}/reference/codelists#currency)."
-network_id,INTEGER,FK,Network ID,
