--- schema/geopackage/table_definitions/wayleaves_fks.csv
@@ -1,3 +1,0 @@
-Column,References
-network_id,FK(`networks`.`id`)
-grantor,FK(`organisations`.`id`)
